fix signed/unsigned warning
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@9423 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/gtfont.cpp
@@ -704,7 +704,7 @@ void gtFont::parseFamily()
 		family = name;
 	else
 		family = name.left(smallestIndex);
-	if (biggestIndex == -1 || biggestIndex >= name.length())
+	if (biggestIndex == -1 || biggestIndex >= (int) name.length())
 		append = "";
 	else
 		append = name.right(name.length() - biggestIndex - 1);
